Move function to multiple testing
