+++ examples/plot_knockoff_aggregation.py
@@ -34,8 +34,8 @@ from hidimstat.knockoffs import (
     model_x_knockoff_bootstrap_quantile,
     model_x_knockoff_pvalue,
 )
+from hidimstat.statistical_tools.multiple_testing import cal_fdp_power
 from hidimstat._utils.scenario import multivariate_1D_simulation_AR
-from hidimstat._utils.utils import cal_fdp_power
 
 plt.rcParams.update({"font.size": 26})
